Trained a multitude of different Models
